prepare using codebase to generate importmap as well
