+++ script/generate-import-map/generate-import-map.js
@@ -8,6 +8,8 @@ generateImportMapForProject(
   [
     getImportMapFromNodeModules({
       projectDirectoryUrl,
+      target: "node",
+      dev: true,
     }),
   ],
   {
