+++ src/internal/package/applyPackageManualOverride.js
