+++ src/internal/package/node-module-resolution.js
@@ -1,6 +1,6 @@
 import { firstOperationMatching } from "@jsenv/cancellation"
 import { urlToRelativeUrl, resolveUrl } from "@jsenv/util"
-import { memoizeAsyncFunctionByUrl } from "./memoizeAsyncFunctionByUrl.js"
+import { memoizeAsyncFunctionByUrl } from "../memoizeAsyncFunctionByUrl.js"
 import { readPackageFile, PACKAGE_NOT_FOUND, PACKAGE_WITH_SYNTAX_ERROR } from "./readPackageFile.js"
 
 export const createFindNodeModulePackage = (packagesManualOverrides) => {
