+++ src/internal/package/readPackageFile.js
