+++ src/internal/package/resolveFile.js
