+++ src/internal/package/resolveNodeModule.js
