+++ src/internal/package/resolvePackageMain.js
