+++ src/internal/package/specifierIsRelative.js
