+++ src/internal/package/visitPackageImports.js
